Make simplify_types more robust
These modifications ensure that all nodes have a .abstract after
the process has run its course. It also converts all constants that
may have values with invalid types to the proper values (e.g. it
will convert a dataclass to a tuple if it is in a constant).

A coerce argument is also added to to_canonical in order to bypass
type checking in cases where e.g. we have a constant with value 1
(type i64) and type f64, which is fine. Ideally it should convert
this value to the right type, but this causes some problems
currently.