Add "quarantine" option to clone
This is likely a temporary change. Regarding the expand_J optimization specifically,
it is necessary to perform copies of graphs, however we only want to copy new graphs
created by the gradient transform which have no type data, and not existing graphs
that may be referred to by the new graphs, but do have type information.

These existing graphs, if they remain, will interfere with the monomorphization step
because they belong to a different manager than the (new) manager used to monomorphize
the gradient graphs.

The quarantine option is a function that will wrap graphs during cloning using some
arbitrary criterion, although the one used in practice is that graphs that have type
information are wrapped. Because of the wrapping they go through inference and
monomorphization unchanged, and then the wrapping can be undone.

That was the simplest way I found to make this work. It is not ideal, but we may
want to try doing inference+monomorphize without the manager, which would obviate
the need for this workaround.