Merge functions with VirtualFunction in amerge
* Merging Possibilities with multiple VirtualFunctions will create
a Possibilities with only one VirtualFunction.
* If any possibility is a VirtualFunction, the inferrer is enlisted
to verify that the other possibilities are compatible: inference is
scheduled to check that the GraphFunction or other has the same
output as the VirtualFunction when given the same args.

This uses a hack of sorts to provide the engine to amerge, which is
to set and read a ContextVar (this is thread/coroutine safe). Didn't
feel like adding an argument, and there may be better ways to achieve
this in the future that could obviate the need for this hack.