Add ForceConstants optimization
Prior to incremental inference, the inferrer/monomorphizer would
replace nodes like make_tuple(1, 2) by a constant containing (1, 2),
but now that there are no renormalize passes, there needs to be an
optimization that does more or less the same thing.

This optimization simply replaces nodes by constants whenever it is
able to build the constant from the abstract field.