Validate functions and add CallValidator
AbstractFunctions must now contain exactly one VirtualFunction at
the end of the pipeline.

CallValidator verifies that the type of f in f(x, y) is consistent
with the arguments and return value. Unfortunately there are still
hiccups that make this check fail, although the tests pass
otherwise. I'm leaving the code around for the time being, though.