Turn off a few tests in test_grad
Due to issues with manager conflicts, it is not possible to use grad
on graphs that have free variables (these graphs can still contain
other graphs). This is something to fix later.

Another test, hof_tup, requires more optimizations to be written in
order to work (otherwise it is not possible to eliminate all
AbstractJTagged nodes).