Adjust a few optimization tests with respect to typing
Optimizations are now in charge of typing the graphs they insert,
so a test that inserted a graph to check that it's properly typed
had to be removed.

I added a new test which unfortunately fails, because we don't
detect that the node inserted by an optimization has the same type
as the original, when its type gets inferred.