Adapt gradient computation to the new inferrer
The expand_J optimization now runs the inferrer and monomorphizer
on the gradient graph in order to make sure that the graphs it
inserts are all property typed and monomorphized.

This causes a few issues with the manager, so to make things simple,
this is all done with a fresh manager. This means imported graphs
using convert should not use the main manager, hence the manage
option I added.

The replace_Jinv_on_graph was also required to replace a graph
by its primal.